chore(oms): move platform models to domain package
