+++ app/db/base.py
@@ -142,6 +142,7 @@ def init_models(
         "app.wms.outbound.models",
         "app.tms.records.models",
         "app.tms.billing.models",
+        "app.oms.platforms.models",
     ):
         for mod in _iter_model_modules_recursive(pkg_name):
             if mod in ex or mod in loaded:
