+++ app/models/__init__.py
@@ -86,18 +86,18 @@ MODEL_SPECS = [
     # ------------------------------------------------------------------
     # 店铺 × 平台接入（OMS 淘宝 / 拼多多 / 京东）
     # ------------------------------------------------------------------
-    ("app.models.taobao_app_config", "TaobaoAppConfig"),
-    ("app.models.pdd_app_config", "PddAppConfig"),
-    ("app.models.jd_app_config", "JdAppConfig"),
-    ("app.models.store_platform_credential", "StorePlatformCredential"),
-    ("app.models.store_platform_connection", "StorePlatformConnection"),
-    ("app.models.taobao_order", "TaobaoOrder"),
-    ("app.models.taobao_order", "TaobaoOrderItem"),
-    ("app.models.pdd_order", "PddOrder"),
-    ("app.models.pdd_order", "PddOrderItem"),
-    ("app.models.jd_order", "JdOrder"),
-    ("app.models.jd_order", "JdOrderItem"),
-    ("app.models.pdd_order_order_mapping", "PddOrderOrderMapping"),
+    ("app.oms.platforms.models.taobao_app_config", "TaobaoAppConfig"),
+    ("app.oms.platforms.models.pdd_app_config", "PddAppConfig"),
+    ("app.oms.platforms.models.jd_app_config", "JdAppConfig"),
+    ("app.oms.platforms.models.store_platform_credential", "StorePlatformCredential"),
+    ("app.oms.platforms.models.store_platform_connection", "StorePlatformConnection"),
+    ("app.oms.platforms.models.taobao_order", "TaobaoOrder"),
+    ("app.oms.platforms.models.taobao_order", "TaobaoOrderItem"),
+    ("app.oms.platforms.models.pdd_order", "PddOrder"),
+    ("app.oms.platforms.models.pdd_order", "PddOrderItem"),
+    ("app.oms.platforms.models.jd_order", "JdOrder"),
+    ("app.oms.platforms.models.jd_order", "JdOrderItem"),
+    ("app.oms.platforms.models.pdd_order_order_mapping", "PddOrderOrderMapping"),
     # ------------------------------------------------------------------
     # 采购系统
     # ------------------------------------------------------------------
