+++ app/oms/platforms/models/__init__.py
@@ -1,0 +1,29 @@
+# app/oms/platforms/models/__init__.py
+# Domain-owned ORM models for OMS platform access and platform order facts.
+
+from app.oms.platforms.models.jd_app_config import JdAppConfig
+from app.oms.platforms.models.jd_order import JdOrder, JdOrderItem
+from app.oms.platforms.models.pdd_app_config import PddAppConfig
+from app.oms.platforms.models.pdd_order import PddOrder, PddOrderItem
+from app.oms.platforms.models.pdd_order_order_mapping import PddOrderOrderMapping
+from app.oms.platforms.models.store_platform_connection import StorePlatformConnection
+from app.oms.platforms.models.store_platform_credential import StorePlatformCredential
+from app.oms.platforms.models.store_token import StoreToken
+from app.oms.platforms.models.taobao_app_config import TaobaoAppConfig
+from app.oms.platforms.models.taobao_order import TaobaoOrder, TaobaoOrderItem
+
+__all__ = [
+    "JdAppConfig",
+    "JdOrder",
+    "JdOrderItem",
+    "PddAppConfig",
+    "PddOrder",
+    "PddOrderItem",
+    "PddOrderOrderMapping",
+    "StorePlatformConnection",
+    "StorePlatformCredential",
+    "StoreToken",
+    "TaobaoAppConfig",
+    "TaobaoOrder",
+    "TaobaoOrderItem",
+]
