+++ app/oms/platforms/models/pdd_app_config.py
@@ -1,4 +1,5 @@
-# app/models/taobao_app_config.py
+# app/oms/platforms/models/pdd_app_config.py
+# Domain move: PDD app config ORM belongs to OMS platform access.
 from __future__ import annotations
 
 from datetime import datetime
@@ -9,13 +10,13 @@ from sqlalchemy.orm import Mapped, mapped_column
 from app.db.base import Base
 
 
-class TaobaoAppConfig(Base):
+class PddAppConfig(Base):
     """
-    淘宝开放平台系统级应用配置表。
+    拼多多开放平台系统级应用配置表。
 
     职责：
-    - 保存 OMS 系统接入淘宝开放平台所需的应用配置
-    - 保存 app_key / app_secret / callback_url / api_base_url / sign_method
+    - 保存 OMS 系统接入拼多多开放平台所需的应用配置
+    - 保存 client_id / client_secret / redirect_uri / api_base_url / sign_method
     - 保存当前是否启用的配置记录
 
     不负责：
@@ -24,11 +25,11 @@ class TaobaoAppConfig(Base):
     - 店铺 identity
     """
 
-    __tablename__ = "taobao_app_configs"
+    __tablename__ = "pdd_app_configs"
 
     __table_args__ = (
         sa.Index(
-            "ix_taobao_app_configs_is_enabled",
+            "ix_pdd_app_configs_is_enabled",
             "is_enabled",
         ),
     )
@@ -39,17 +40,17 @@ class TaobaoAppConfig(Base):
         autoincrement=True,
     )
 
-    app_key: Mapped[str] = mapped_column(
+    client_id: Mapped[str] = mapped_column(
         sa.String(128),
         nullable=False,
     )
 
-    app_secret: Mapped[str] = mapped_column(
+    client_secret: Mapped[str] = mapped_column(
         sa.Text,
         nullable=False,
     )
 
-    callback_url: Mapped[str] = mapped_column(
+    redirect_uri: Mapped[str] = mapped_column(
         sa.String(512),
         nullable=False,
     )
@@ -57,7 +58,7 @@ class TaobaoAppConfig(Base):
     api_base_url: Mapped[str] = mapped_column(
         sa.String(255),
         nullable=False,
-        server_default=sa.text("'https://eco.taobao.com/router/rest'"),
+        server_default=sa.text("'https://gw-api.pinduoduo.com/api/router'"),
     )
 
     sign_method: Mapped[str] = mapped_column(
