+++ app/oms/platforms/models/taobao_order.py
@@ -1,3 +1,5 @@
+# app/oms/platforms/models/taobao_order.py
+# Domain move: Taobao order fact ORM belongs to OMS platform order ledger.
 from __future__ import annotations
 
 from datetime import datetime
@@ -11,43 +13,43 @@ from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.db.base import Base
 
 
-class JdOrder(Base):
+class TaobaoOrder(Base):
     """
-    京东平台订单头事实表（jd_orders）。
+    淘宝平台订单头事实表（taobao_orders）。
 
     职责：
-    - 保存 JD 平台订单头原生事实字段
+    - 保存淘宝平台订单头原生事实字段
     - 保存收件信息快照
-    - 保存 JD 摘要 / 详情原始 payload
+    - 保存淘宝摘要 / 详情原始 payload
 
     不负责：
     - 保存内部桥接 / 匹配 / 准入 / 建单状态
     - 替代内部 orders 主表
     """
 
-    __tablename__ = "jd_orders"
+    __tablename__ = "taobao_orders"
 
     __table_args__ = (
         sa.UniqueConstraint(
             "store_id",
-            "order_id",
-            name="uq_jd_orders_store_order_id",
+            "tid",
+            name="uq_taobao_orders_store_tid",
         ),
         sa.Index(
-            "ix_jd_orders_store_id",
+            "ix_taobao_orders_store_id",
             "store_id",
         ),
         sa.Index(
-            "ix_jd_orders_order_state",
-            "order_state",
+            "ix_taobao_orders_status",
+            "status",
         ),
         sa.Index(
-            "ix_jd_orders_order_start_time",
-            "order_start_time",
+            "ix_taobao_orders_created",
+            "created",
         ),
         sa.Index(
-            "ix_jd_orders_modified",
-            "modified",
+            "ix_taobao_orders_pay_time",
+            "pay_time",
         ),
     )
 
@@ -64,130 +66,148 @@ class JdOrder(Base):
         comment="OMS 店铺 id（stores.id）",
     )
 
-    order_id: Mapped[str] = mapped_column(
+    tid: Mapped[str] = mapped_column(
         sa.String(64),
         nullable=False,
-        comment="京东主订单号 order_id",
+        comment="淘宝主订单号 tid",
     )
 
-    vender_id: Mapped[str | None] = mapped_column(
+    status: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
-        comment="京东商家 id / vender_id",
+        comment="淘宝原始交易状态",
     )
 
-    order_type: Mapped[str | None] = mapped_column(
+    type: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
-        comment="京东原始订单类型",
+        comment="淘宝原始交易类型",
     )
 
-    order_state: Mapped[str | None] = mapped_column(
-        sa.String(64),
-        nullable=True,
-        comment="京东原始订单状态",
-    )
-
-    buyer_pin: Mapped[str | None] = mapped_column(
+    buyer_nick: Mapped[str | None] = mapped_column(
         sa.String(128),
         nullable=True,
-        comment="买家标识 buyer_pin",
+        comment="买家昵称",
     )
 
-    consignee_name: Mapped[str | None] = mapped_column(
+    buyer_open_uid: Mapped[str | None] = mapped_column(
+        sa.String(128),
+        nullable=True,
+        comment="买家 OpenUID",
+    )
+
+    receiver_name: Mapped[str | None] = mapped_column(
         sa.String(128),
         nullable=True,
         comment="收件人姓名",
     )
 
-    consignee_mobile: Mapped[str | None] = mapped_column(
+    receiver_mobile: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
         comment="收件人手机号",
     )
 
-    consignee_phone: Mapped[str | None] = mapped_column(
+    receiver_phone: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
         comment="收件人电话",
     )
 
-    consignee_province: Mapped[str | None] = mapped_column(
+    receiver_state: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
         comment="收件省",
     )
 
-    consignee_city: Mapped[str | None] = mapped_column(
+    receiver_city: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
         comment="收件市",
     )
 
-    consignee_county: Mapped[str | None] = mapped_column(
+    receiver_district: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
         comment="收件区/县",
     )
 
-    consignee_town: Mapped[str | None] = mapped_column(
+    receiver_town: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
         comment="收件街道/镇",
     )
 
-    consignee_address: Mapped[str | None] = mapped_column(
+    receiver_address: Mapped[str | None] = mapped_column(
         sa.String(512),
         nullable=True,
         comment="收件详细地址",
     )
 
-    order_remark: Mapped[str | None] = mapped_column(
-        sa.Text,
+    receiver_zip: Mapped[str | None] = mapped_column(
+        sa.String(32),
         nullable=True,
-        comment="订单备注 / 买家备注",
+        comment="收件邮编",
     )
 
-    seller_remark: Mapped[str | None] = mapped_column(
+    buyer_memo: Mapped[str | None] = mapped_column(
+        sa.Text,
+        nullable=True,
+        comment="买家备注",
+    )
+
+    buyer_message: Mapped[str | None] = mapped_column(
+        sa.Text,
+        nullable=True,
+        comment="买家留言/附言",
+    )
+
+    seller_memo: Mapped[str | None] = mapped_column(
         sa.Text,
         nullable=True,
         comment="卖家备注",
     )
 
-    order_total_price: Mapped[Decimal | None] = mapped_column(
+    seller_flag: Mapped[int | None] = mapped_column(
+        sa.Integer,
+        nullable=True,
+        comment="卖家备注旗帜",
+    )
+
+    payment: Mapped[Decimal | None] = mapped_column(
         sa.Numeric(14, 2),
         nullable=True,
-        comment="订单总金额（元）",
+        comment="实付金额（元）",
     )
 
-    order_seller_price: Mapped[Decimal | None] = mapped_column(
+    total_fee: Mapped[Decimal | None] = mapped_column(
         sa.Numeric(14, 2),
         nullable=True,
-        comment="商家应收金额（元）",
+        comment="应付金额（元）",
     )
 
-    freight_price: Mapped[Decimal | None] = mapped_column(
+    post_fee: Mapped[Decimal | None] = mapped_column(
         sa.Numeric(14, 2),
         nullable=True,
-        comment="运费金额（元）",
+        comment="邮费（元）",
     )
 
-    payment_confirm: Mapped[str | None] = mapped_column(
-        sa.String(32),
+    coupon_fee: Mapped[Decimal | None] = mapped_column(
+        sa.Numeric(14, 2),
         nullable=True,
-        comment="付款确认状态",
+        comment="优惠券金额（元）",
     )
 
-    order_start_time: Mapped[datetime | None] = mapped_column(
+    created: Mapped[datetime | None] = mapped_column(
         sa.DateTime(timezone=True),
         nullable=True,
-        comment="下单时间 / 订单开始时间",
+        comment="交易创建时间",
     )
 
-    order_end_time: Mapped[datetime | None] = mapped_column(
+    pay_time: Mapped[datetime | None] = mapped_column(
         sa.DateTime(timezone=True),
         nullable=True,
-        comment="订单结束时间",
+        comment="付款时间",
     )
 
     modified: Mapped[datetime | None] = mapped_column(
@@ -199,13 +219,13 @@ class JdOrder(Base):
     raw_summary_payload: Mapped[dict | None] = mapped_column(
         JSONB,
         nullable=True,
-        comment="JD 摘要接口原始 payload",
+        comment="淘宝摘要接口原始 payload",
     )
 
     raw_detail_payload: Mapped[dict | None] = mapped_column(
         JSONB,
         nullable=True,
-        comment="JD 详情接口原始 payload",
+        comment="淘宝详情接口原始 payload",
     )
 
     pulled_at: Mapped[datetime] = mapped_column(
@@ -237,8 +257,8 @@ class JdOrder(Base):
         comment="记录更新时间",
     )
 
-    items: Mapped[List["JdOrderItem"]] = relationship(
-        "JdOrderItem",
+    items: Mapped[List["TaobaoOrderItem"]] = relationship(
+        "TaobaoOrderItem",
         back_populates="order",
         lazy="selectin",
         cascade="all, delete-orphan",
@@ -246,39 +266,46 @@ class JdOrder(Base):
 
     def __repr__(self) -> str:
         return (
-            f"<JdOrder id={self.id} store_id={self.store_id} "
-            f"order_id={self.order_id} state={self.order_state}>"
+            f"<TaobaoOrder id={self.id} store_id={self.store_id} "
+            f"tid={self.tid} status={self.status}>"
         )
 
 
-class JdOrderItem(Base):
+class TaobaoOrderItem(Base):
     """
-    京东平台订单行事实表（jd_order_items）。
+    淘宝平台子订单事实表（taobao_order_items）。
 
     职责：
-    - 保存 JD 平台原始商品行字段
-    - 保存 JD 订单行原始 payload
+    - 保存淘宝子订单原生商品行字段
+    - 保存子订单原始 payload
     """
 
-    __tablename__ = "jd_order_items"
+    __tablename__ = "taobao_order_items"
 
     __table_args__ = (
         sa.UniqueConstraint(
-            "jd_order_id",
-            "sku_id",
-            "ware_id",
-            name="uq_jd_order_items_order_sku_ware",
+            "taobao_order_id",
+            "oid",
+            name="uq_taobao_order_items_order_oid",
         ),
         sa.Index(
-            "ix_jd_order_items_jd_order_id",
-            "jd_order_id",
+            "ix_taobao_order_items_taobao_order_id",
+            "taobao_order_id",
         ),
         sa.Index(
-            "ix_jd_order_items_order_id",
-            "order_id",
+            "ix_taobao_order_items_tid",
+            "tid",
         ),
         sa.Index(
-            "ix_jd_order_items_outer_sku_id",
+            "ix_taobao_order_items_oid",
+            "oid",
+        ),
+        sa.Index(
+            "ix_taobao_order_items_outer_iid",
+            "outer_iid",
+        ),
+        sa.Index(
+            "ix_taobao_order_items_outer_sku_id",
             "outer_sku_id",
         ),
     )
@@ -289,23 +316,41 @@ class JdOrderItem(Base):
         autoincrement=True,
     )
 
-    jd_order_id: Mapped[int] = mapped_column(
+    taobao_order_id: Mapped[int] = mapped_column(
         sa.BigInteger,
-        sa.ForeignKey("jd_orders.id", ondelete="CASCADE"),
+        sa.ForeignKey("taobao_orders.id", ondelete="CASCADE"),
         nullable=False,
-        comment="所属 JD 订单头 id",
+        comment="所属淘宝订单头 id",
     )
 
-    order_id: Mapped[str] = mapped_column(
+    tid: Mapped[str] = mapped_column(
         sa.String(64),
         nullable=False,
-        comment="京东主订单号（冗余保存）",
+        comment="淘宝主订单号 tid（冗余保存）",
+    )
+
+    oid: Mapped[str] = mapped_column(
+        sa.String(64),
+        nullable=False,
+        comment="淘宝子订单号 oid",
+    )
+
+    num_iid: Mapped[str | None] = mapped_column(
+        sa.String(64),
+        nullable=True,
+        comment="商品数字 ID",
     )
 
     sku_id: Mapped[str | None] = mapped_column(
         sa.String(64),
         nullable=True,
-        comment="京东 sku_id",
+        comment="SKU ID",
+    )
+
+    outer_iid: Mapped[str | None] = mapped_column(
+        sa.String(128),
+        nullable=True,
+        comment="商家外部商品编码",
     )
 
     outer_sku_id: Mapped[str | None] = mapped_column(
@@ -314,47 +359,47 @@ class JdOrderItem(Base):
         comment="商家外部 SKU 编码",
     )
 
-    ware_id: Mapped[str | None] = mapped_column(
-        sa.String(64),
-        nullable=True,
-        comment="京东商品 ware_id",
-    )
-
-    item_name: Mapped[str | None] = mapped_column(
+    title: Mapped[str | None] = mapped_column(
         sa.String(255),
         nullable=True,
-        comment="商品名称",
+        comment="商品标题",
     )
 
-    item_total: Mapped[int] = mapped_column(
+    price: Mapped[Decimal | None] = mapped_column(
+        sa.Numeric(14, 2),
+        nullable=True,
+        comment="单价（元）",
+    )
+
+    num: Mapped[int] = mapped_column(
         sa.Integer,
         nullable=False,
         server_default="0",
         comment="购买数量",
     )
 
-    item_price: Mapped[Decimal | None] = mapped_column(
+    payment: Mapped[Decimal | None] = mapped_column(
         sa.Numeric(14, 2),
         nullable=True,
-        comment="单价（元）",
+        comment="子订单实付金额（元）",
     )
 
-    sku_name: Mapped[str | None] = mapped_column(
+    total_fee: Mapped[Decimal | None] = mapped_column(
+        sa.Numeric(14, 2),
+        nullable=True,
+        comment="子订单应付金额（元）",
+    )
+
+    sku_properties_name: Mapped[str | None] = mapped_column(
         sa.Text,
         nullable=True,
-        comment="SKU 规格描述",
-    )
-
-    gift_point: Mapped[int | None] = mapped_column(
-        sa.Integer,
-        nullable=True,
-        comment="是否赠品标识 / gift_point",
+        comment="SKU 属性名串",
     )
 
     raw_item_payload: Mapped[dict | None] = mapped_column(
         JSONB,
         nullable=True,
-        comment="JD 行原始 payload",
+        comment="淘宝子订单原始 payload",
     )
 
     created_at: Mapped[datetime] = mapped_column(
@@ -372,14 +417,14 @@ class JdOrderItem(Base):
         comment="记录更新时间",
     )
 
-    order: Mapped["JdOrder"] = relationship(
-        "JdOrder",
+    order: Mapped["TaobaoOrder"] = relationship(
+        "TaobaoOrder",
         back_populates="items",
         lazy="selectin",
     )
 
     def __repr__(self) -> str:
         return (
-            f"<JdOrderItem id={self.id} jd_order_id={self.jd_order_id} "
-            f"sku_id={self.sku_id} qty={self.item_total}>"
+            f"<TaobaoOrderItem id={self.id} taobao_order_id={self.taobao_order_id} "
+            f"oid={self.oid} qty={self.num}>"
         )
