+++ app/oms/platforms/taobao/repository.py
@@ -8,9 +8,9 @@ from typing import Optional
 import sqlalchemy as sa
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.models.store_platform_connection import StorePlatformConnection
-from app.models.store_platform_credential import StorePlatformCredential
-from app.models.taobao_app_config import TaobaoAppConfig
+from app.oms.platforms.models.store_platform_connection import StorePlatformConnection
+from app.oms.platforms.models.store_platform_credential import StorePlatformCredential
+from app.oms.platforms.models.taobao_app_config import TaobaoAppConfig
 
 
 def _norm_platform(platform: str) -> str:
