+++ tests/api/test_jd_orders_contract.py
@@ -8,7 +8,7 @@ from sqlalchemy import text
 
 from app.user.deps.auth import get_current_user
 from app.main import app
-from app.models.jd_order import JdOrder, JdOrderItem
+from app.oms.platforms.models.jd_order import JdOrder, JdOrderItem
 import app.oms.platforms.jd.router_orders as jd_router_orders
