+++ tests/api/test_pdd_connection_and_pull_contract.py
@@ -7,9 +7,9 @@ from sqlalchemy import text
 
 from app.user.deps.auth import get_current_user
 from app.main import app
-from app.models.pdd_app_config import PddAppConfig
-from app.models.store_platform_connection import StorePlatformConnection
-from app.models.store_platform_credential import StorePlatformCredential
+from app.oms.platforms.models.pdd_app_config import PddAppConfig
+from app.oms.platforms.models.store_platform_connection import StorePlatformConnection
+from app.oms.platforms.models.store_platform_credential import StorePlatformCredential
 
 
 class _TestUser:
